+++ packages/component/src/SpeechServices/SpeechSDK.js
@@ -1,5 +1,3 @@
-// TODO: [P1] This need to be fixed with a peer depending package, instead of a burnt in bundle.
+import * as SDK from 'microsoft-cognitiveservices-speech-sdk';
 
-import './external/microsoft.cognitiveservices.speech.sdk.bundle';
-
-export default window.SpeechSDK
+export default SDK